+++ package.json
@@ -5,10 +5,10 @@
 		"dev": "esbuild --bundle src/main.js --outfile=www/main.js --servedir=www",
 		"zip": "npm run build && mkdir -p dist && zip -r dist/game.zip www -x \"**/.DS_Store\""
 	},
-	"dependencies": {
-		"kaboom": "^2000.2.10"
-	},
 	"devDependencies": {
 		"esbuild": "^0.17.18"
+	},
+	"dependencies": {
+		"kaboom": "^2000.2.10"
 	}
 }
